+++ Runtime/Services/Resource/Handler/YooAssetHandler.Records.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using UnityEngine;
-using UObject = UnityEngine.Object;
 
 namespace Moirai.Atropos.Resource
 {